+++ src/Equation.py
